# api-specification-singulary.yaml
openapi: 3.0.3
info:
  title: Singulary API
  description: |
    API completa para o aplicativo Singulary - Ressonador Quântico Pessoal (RQP).
    
    Esta API fornece endpoints para gerenciamento de usuários, presets de frequências,
    sessões de meditação/terapia sonora e integração futura com hardware.
    
    ## Funcionalidades Principais
    - Autenticação JWT
    - Gerenciamento de presets (padrão e personalizados)
    - Registro e análise de sessões
    - Exportação de dados em CSV
    - Estatísticas e relatórios
    - Preparação para integração com hardware
    
    ## Segurança
    - Autenticação via JWT tokens
    - Validação de dados de entrada
    - Rate limiting (implementação futura)
    - CORS configurado para frontend
    
    ## Versioning
    Esta é a versão 1.0 da API, compatível com o MVP do Singulary.
    
  version: 1.0.0
  contact:
    name: Equipe Singulary
    email: api@singulary.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Servidor de desenvolvimento local
  - url: https://api.singulary.app/v1
    description: Servidor de produção (futuro)
...
paths:
...
  /auth/register:
    post:
      tags:
        - auth
      summary: Registrar novo usuário
      description: Criar uma nova conta de usuário no sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                name:
                  type: string
                  example: João Silva
                  description: Nome completo do usuário
                email:
                  type: string
                  format: email
                  example: joao@example.com
                  description: Email único do usuário
                password:
                  type: string
                  minLength: 6
                  example: senha123
                  description: Senha com pelo menos 6 caracteres
                consent:
                  type: boolean
                  example: true
                  description: Consentimento para termos de uso
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Usuário criado com sucesso
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email já está em uso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - auth
      summary: Fazer login
      description: Autenticar usuário e obter token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: joao@example.com
                password:
                  type: string
                  example: senha123
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login realizado com sucesso
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/verify:
    post:
      tags:
        - auth
      summary: Verificar token JWT
      description: Validar se o token JWT é válido e obter informações do usuário
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token válido
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags:
        - auth
      summary: Renovar token JWT
      description: Gerar um novo token JWT válido
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token renovado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token renovado com sucesso
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/change-password:
    post:
      tags:
        - auth
      summary: Alterar senha
      description: Alterar a senha do usuário autenticado
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  example: senhaAtual123
                newPassword:
                  type: string
                  minLength: 6
                  example: novaSenha456
      responses:
        '200':
          description: Senha alterada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Senha alterada com sucesso
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # PRESETS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtido através do endpoint de login.
        Formato: `Bearer <token>`
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: user-abc123def456
          description: ID único do usuário
        name:
          type: string
          example: João Silva
          description: Nome completo do usuário
        email:
          type: string
          format: email
          example: joao@example.com
          description: Email do usuário
        createdAt:
          type: string
          format: date-time
          example: 2025-08-01T10:00:00Z
          description: Data de criação da conta
        updatedAt:
          type: string
          format: date-time
          example: 2025-08-10T15:30:00Z
          description: Data da última atualização
        lastLogin:
          type: string
          format: date-time
          example: 2025-08-10T14:00:00Z
          description: Data do último login
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Mensagem de erro
          description: Descrição do erro
        code:
          type: string
          example: INVALID_INPUT
          description: Código do erro (opcional)
        details:
          type: object
          description: Detalhes adicionais do erro (opcional)
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Dados de entrada inválidos
            code: INVALID_INPUT
    Unauthorized:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Token inválido ou expirado
            code: UNAUTHORIZED
...